Unite jellies and water elementals under M_AMORPHOUS, tweak other checks
There's been some consternation for a while that water elementals and
elemental wellsprings, as entities composed of liquids, should have the
same immunities that jellies get (constriction, webs, nets as of 4fe0c3d).
There have been a few requests to make them insubstantial (c.f. PR #3032),
but I would prefer to instead unite them with jellies in a new flag that
can reduce the number of jelly genus checks, instead of granting the
water-formed enemies immunity to petrification or implying that a monster
which can entirely engulf a player also has no actual substance.

This commit adds the M_AMORPHOUS flag, shared by both jellies and water
elemental / wellsprings. This gives the latter web, net, constriction
immunity, and Shatter resistance as the former does, and hopefully
should feel a bit more logically consistent while cleaning up a weird
check scattered across the code base. This commit additionally teaches
the monster information bots to display insubstantial and amorphous
flags, as is useful for offhandedly checking net and BVC immunity.

(This commit also removes the barbs immunity from jellies. If they'd be
immune to barbs because they can force the barbs out of themselves, then
the same can be said of all sorts of other physical projectile states,
like Frozen. The committer of dd4bd0d agrees with doing this.)

FIXME: Borgjnor's Vile Clutch always listed the hit chance as 100% for
insubstantial enemies, and continues to display it for amorphous enemies.
It's a very silly targetter since it's always 100% or 0%, but it should
show the latter for those immune to it.