feat: Add monster intelligence to clua
Monster intelligence (e.g. "Human" or "Mindless") was missing from clua
even though it's present in the UI. This is a minimal implementation
that simply returns the standard description string of a monster's
intelligence. This is technically a feature, which we wouldn't normally
add to stable/tournament, but it's only related to UI and arguably a bug
that it was missing. It's also very minimal and has been tested locally.