Marginally nerf ball pythons
After the major constriction rewrite of 9adf0d1, ball python have finally
started to successfully tutorialize on the existence and power of
constriction for D:1. This is good! It means there's a monster with a
gimmick that'll take a fair ways longer to be seen again, that can allow
experimenting with the gimmick much earlier, and which stands out amongst
other D:1 and D:2 enemies in its support + damage role.

However, they're also a noticeably increased threat on D:1, which doesn't
really need the bump upwards- between the two tournaments, they moved from
the 43rd killer to the 13th. A good promotion, but possibly a bit much for
one.

As such, I'm nudging down their non-constriction damage and their EV very
slightly, which should make it easier to kill them when there's other
enemies around, but which should also still let their newly finally
threatening constriction be somewhat noticeable if they're not focused down
on.