Let player berserk give immunity to fear (DracoOmega)
Berserk monsters are immune to fear when berserked (as is obvious with
early black bears), but players aren't (and a comment has been wondering
about this discrepancy since c6d2f85). Cause Fear is a relatively rare
effect only used on about four monsters, so it's not too significant a
monster nerf / berserk buff (and those monsters all either meaningfully
summon or debuff from a distance, so there's still some danger), but it
adds a little more consistency in a relatively player-generous fashion.

(Also of high importance: Crawl doesn't show the fear aura that well when
one is also berserk, and it'd be a massive headache to try to make it do
so. Happy dual solutions.)