Mark milestones when buying runes (grothendieck)
Buying items in shops always creates a note and then flags the item to
avoid creating duplicate notes, but that flag then also caused milestone
recording to be skipped. Historically this wasn't an issue because the only
items that need milestones are runes/gems/the Orb but at least one of those
is now purchasable in a rare case!