Make exploding sheep properly flee from the player (Lici)
The dream sheep spawned by kb_xom_exploding_sheep are generated permanently
afraid, with the presumable intent that they run away from the player,
however they aren't generated being *aware* of the player, which means that
they usually have an internal target of (0, 0) and they try to run away
from that instead of the player (ie: they just move south-east in general,
regardless of where the player is).

If any other monster shouted or the player otherwise got the sheep's
attention somehow, they would start fleeing in the proper direction, but it
could still sometimes be a rude surprise to a low XL player. So let's try
to get them moving in the right direction to start with.