Preserve bribe enchantments for reviving bennu
Previously, if a bennu was friendly-bribed in their first life, it
would revive as a permanent friendly ally that could follow the player
everywhere. If it was temporarily neutral-bribed, it would revive
as a permanent neutral monster.

Fix this issue by preserving friendly and neutral bribe enchantments
upon revival.

Sidenote: currently unbribed bennu get a second chance to become bribed
upon revival - is this good behaviour?

Resolves #4026.