Don't auto-hide currently unmemorisable spells when picking up books
Previously, if you worshipped Okawaru (for example), spells which
could not be memorised would be silently added to your library and also
hidden in your library.

Spells that can't be memorised do not appear on the spell library
interface, so this had no actual effect as long as the spell could not
be memorised. But, say, if you then abandoned Okawaru, you would
find all their previously unmemorisable spells hidden in the library
(so that, to unhide and memorise them, you would have to visit the
"Show" screen).

In practice, this meant many players would just assume the spell was
lost forever.

Resolves #3994.