Don't gain experience from killing allied apostles
Previously it was possible to kill your apostles with e.g. scrolls of
poison, not incur wrath, and gain experience from the kill.

I can't really imagine players ever wanting to take advantage of this,
but nonetheless it goes against the spirit of Beogh so let's prevent
apostles from giving experience.