Fix a rare crash with Dith shadow shot
If the player made a ranged attack which killed its primary target, and
also triggered a shadow shot while the player shadow was not already out
*and* the player was both invisible and could not see invisible, the game
would crash. If any of these things was not true, it worked properly.

(The issue stemmed from fuzzing for beams near invisible players and how
MID_PLAYER was used as the source of the tracer for determining what other
target the shadow should shoot at, if it wasn't deployed yet. But
MID_PLAYER_SHADOW_DUMMY was added for the specific purpose of simulating
a player shadow that doesn't exist yet, so use it instead here. To be
honest, I'm not sure why it wasn't already.)

(cherry picked from commit 4012fd3d9bc679534802235d590e624e949f8869)