New Pan demon: nekomata
Pandemonium could always do with some more variety beyond the demons one
sees shared between summoning from the rest of a 3-rune game, the demonic
vaults before Lair branches + entry vaults Depths, and the demons shared
with the Abyss and the Hells. Demonspawn enemies and Pan lords themselves
reasonably help with this, but there's plenty of hypothetical space across
multiple glyphs to fit in more monsters, especially in the danger tier
that demonspawn try to fulfill between tier 2 and tier 1 demons.

This new feline demon of trickery and disdain, a white "R" in console, is
meant to try out a strange conjoining of archetypes- it is a ranged weapon
threat _and_ a summoner, simultaneously. In the former aspect, it always
comes with a branded ranged launcher (of flaming or draining, to reflect
its arson and necromantic mythological capacities), plus Repel Missiles
(as we never used in extended) and a demon blade (of either brand) as a
fallback. In the latter aspect, it has a high chance to cast a new
conjurations spell- "Phantom Blitz", a single-target projectile that deals
damage and summons a copy of itself where the projectile stops. The summon
cap for the spell is only 2, but each of the clones have their own summon
cap for the spell, which should be memorably terrifying if let out of
control.

This thus has the unique, cruel property of being even stronger than
previous monsters that summon-on-hit by instead summoning from a distance
with a conjuration, and works as a new iteration on the illusion mechanics
of other eastern demons used in Crawl already. In its placement, it joins
Pan and ranged Zig floors, gets to appear as a mage in Lom Lobon's floor
and in a few guaranteed demonic rune vaults, plus regular Pan floor
monster set generation / orb run sightings. Very high level Xom
worshippers may also see them rarely- a summoning demon that can get out
of control feels entirely appropriate for Xom.

Their tile is a modification of Sastreii's current rakshasa tile (with
twin tails pulled off of Denzi's felid tiles). The current rakshasa tile
is far too cat-like to be any mythologically accurate- a proper rakshasa
may be somewhat beastial, but is hardly a full anthromorphic design, and
also usually somewhat deformed. (Blame D&D for the tiger associations.)
The previously-unused tile by meckryl still emphasizes ornate accessories
while fitting a more fangsome and asymmetric appearance, and is now used
for the rakshasa earlier in the game instead. The Phantom Blitz spell icon
placeholder simply modifies ontoclasm and roctavian's Searing Ray icon.