Revive ancient pizza memes (pianoman523)
Add a new cache of meat tile by pianoman523 that clearly is Just A Pizza
with slight extras. Since caches of meat are meant to convey a wider
variety of meal options beyond pizza itself, the odds of seeing it out in
the wild are relatively low compared to the current placeholder sausage +
jerky tile, but it _can_ be specified by a vault, as has now been done for
a few old vaults and with the revival of one old food vault.