A wide variety of new tiles (CanOfWorms)
New Tiles for the Robe of Augmentation, Firestarter, the rarely-seen base
forms for shapeshifter monsters and glowing shapeshifters, and living
spells now in newly different shades depending on the spell they'll cast.

There's also a new a great orb of eyes tile that I can't help but see a
face in, and have thus mirrored to face left (as every single monster
besides the one literally-sleepwalking monster does). We'll see how it
feels in practice.