Make the Robe of Folly actually a dull grey (#3787)
The 16-year-old rltiles tile for the Robe of Folly has never actually
matched its description as "a dull grey robe", a description roughly
equally old. With a new fanciful Robe of Augmentation tile replacing a
previously grey tile, I've quickly tweaked that old Augmentation tile
to be both a dull grey and stained like the old Robe of Folly tile was.

Closes #3787.