Properly contain a runed door (#4040)
minmay_the_grid_triangle_9x9_subtriangle didn't account for subvault
usages that would connect in the backmost part- I've had to squish in the
vault to make sure it's sealed in. This sometimes means that there's no
actual monsters behind a runed door, but the vault creator has a deep
antipathy against NSUBST and I'd need to review the entire rest of this
file if I was going to make that more consistent anyway.

Closes #4040.