Remove the 'god' argument from most player summoning spells
Despite being in the method signature of many, many spells, this argument
was almost never used, and in the rare situations where it *was* used,
mostly never mattered.

It set the god of the summoned monster to a given value, but monster gods
(especially of player-friendly summons) aren't displayed to the player in
nearly any situation outside of the cast messages of priests (which none of
these spells could even create).

Despite several comments in spl-summoning.cc saying otherwise, this
argument was not the 'god of the caster' and in fact one of the only ways
it could be anything other than GOD_NO_GOD for many spells was when Xom
cast the spell through the player, which would make the resulting monster
worship Xom (not that the player could tell the difference).

Approximately the only functional changes are:
 -If Xom makes the player cast Summon Horrible Things, the player has the
  normal 1-in-5 chance to lose int instead of 0. (This very minor downside
  hasn't been able to be fatal in ages, and double-sided Xom effects are
  something Xom is already full of.)
 -Summons created by player spell cast by Xom will no longer reduce tension
  by an exaggerated amount. Ones created by Xom code directly still will
  (though this may be removed in a future commit)