Refactor backend handling of summons / created monsters (Part 1)
Crawl's handling of summoned and not-technically-summoned-but-close
monsters is a complicated and somewhat untidy thing that makes it harder to
generalize behavior between conventional summons and other monsters that
share many of their properties, without considerable special cases and
bespoke code in various places. For instance, monsters created by the
player can use one of *4* different enchantments for their timer, which can
result in those monsters expiring with with 4 different killer_types (but
not even on a 1-to-1 basis with which timer is used!) that results in
sometimes-unclear differences in behavior. It's also needlessly verbose to
ask if a given monster was created by a given spell, easy to forget to
check for all 4 different ways a monster can be temporary here, and many
other issues that have resulted in various bugs over time (several of which
were outstanding at the time of this commit.)

This is an attempt to significantly clean up and better-document this
entire system, as well as make it easier to interact with and fix a number
of bugs in the process.

Major changes:

-ENCH_ABJ, ENCH_FAKE_ABJURATION, and ENCH_SHORT_LIVED are all folded into
 a single timer: ENCH_SUMMON_TIMER. (Save-compat code will convert
 deprecated enchantments into the new standard timer).

-Rather than differentiate summon behavior by which timer is used, two
 monster flags are added (MF_ACTUAL_SUMMON and MF_PERSISTS) to cover the
 most clear and actually-meaningful differences that already existed.
 -MF_ACTUAL_SUMMON indicates the a monster is abjurable, will produce a
  poof message (and possibly clouds) when it dies - regardless of whether
  this was by timeout or being directly killed - and will not produce other
  death effects (like simulacrum freezing clouds) when it times out.
 -MF_PERSISTS indicates that a monster should not be removed when its
  creator dies (ie: the current behavior for ball lightning and prisms).
 Both of these flags can be easily set by bool parameters to
 mgen_data::set_summoned() (which has a new method signature). Otherwise,
 all summons follow the same set of behaviors.

-monster::is_summoned() will now return true for any sort of 'temporary'
 created monster, regardless of specifically its type. It no longer uses
 outparams.
-monster::is_summoned_by() allows easier querying of summon type (and
 summoner, if desired).

-Summon durations specified in mgen_data now use auts instead of
 'abjuration degree'. The latter was coarse-grained in a way that made it
 problematic to work with for effects that wanted to scale duration with
 power (sometimes resulting in durations that could vary randomly between
 8-200+ turns!), but also: durations for things like ball lightning were
 too short to fit in even the smallest degree, so this change was strictly
 necessary
 -summ_dur(degree) provides a short-hand to emulate the old duration
  behavior, where desired. There are places this ought to be changed, but I
  have not yet done a thorough review of them, and summon duration of most
  things should generally remain unchanged by this commit.
 -The duration of things that used a default ENCH_SHORT_LIVED is adjusted
  slightly (instead of 2.4 - 8.4 aut, they use 4-7)

-Planned-but-not-fully-implemented: All deaths due to summon timers
 expiring will use KILL_TIMEOUT rather than the mishmash of things they
 currently do.

(Not only will this refactoring make it easier to add forgecraft creations
with less special-cased code, but ideally the UI will be slightly improved
by showing the summon dot icon on exactly those monsters which will vanish
with their creator and never on those which will not - removing some
ambiguity that players have complained about in past.)

Note: some changes regarding proper tidiness and unification of killer_type
to make all of the above statements work properly, will be done in
subsequent commits.