New tengu tiles (Sastreii)
Wonderfully sharp tilework as is usual.

(I quickly threw together a new Shadowslip base tengu tile, since it was
missing from the submission and those are relatively simple tiles to make.
Maybe something else besides the tengu conjurer can get a goofy wizard hat
that doesn't blend in so much with the tile in the future?)