Add M_UNSTABLE flag to handle removing some monsters on level change
Several monsters do not play well with how 'catchup' turns work in Crawl
upon returning to a previous floor - hellfire mortar fires multiple times
in a row, ball lightning blow up near where they were the last time the
player was there, even if hundreds of turns have passed.

Removing a few of these upon level exit was already handled, but some were
not, and this generalizes all that by adding a monster flag for 'always
remove summoned instances of this monster when the player leaves the level
they're on'.

(Note: this is *not* already handled by the code to remove player summons
upon leaving a level, since this is an issue for hostile ones as well.)