Rename KILL_MISC to KILL_NON_ACTOR
This better-communicates the largest use of it (direct damage done by
things which are neither the player nor monsters, such as neutral clouds
or some god actions).

It is still used for a few things that arguably don't fit under this
definition, but some of them will be moved out again shortly.