New monster buff: doubled vigour
One troublesome part of making more monster designs for sufficient variety
in several remaining branches is the lack of straightforward buffs
available for monsters to apply to one another. (We _are_ using Battlecry
or Might Other on eight different monsters at this point, as an obvious
example.) It's not the most pressing matter, but additional buffs of
reasonably straightforward and reusable capacities would help.

This new doubled vigour buff extends relatively simply on another common
status effect- berserk, which multiplies both melee damage and hp on
monsters by 150%. This rare buff doesn't affect melee damage, but instead
further increases health by 200% instead. Most uses of hostile healing
beyond Major Healing (on two extended uniques) and innate regeneration
don't really stand out that much, while this should be dramatic enough to
notice. (Since it'll still rescale dealt damage after the status wears off,
it also shouldn't be much more annoying in giving a monster longevity than
berserk itself does). Admittedly, this buff should be used extremely
sparingly as a result, and the current default duration is only as long as
Berserk is.

While this is mostly meant for an Elyvilon-themed monster in an upcoming
commit (thus a name corresponding with the Divine Vigour invocation),
this commit also slips it into the Call of Chaos list (where it can
replace the extremely weak effect of agility to instead give the list
something of high, brief notability besides haste, and thus actually
be a dramatically relevant spell when cast) and onto Xom's hyper-buffing
an ally action (where it can help the earlier xp end of possible provided
allies not immediately die in various situations). Another home may show
up for it in the full scope of time- but for now, it's actually relatively
simple in-code, compared to a lot of other effects, so it shouldn't hurt
any more than Agility somehow still sticking around does.