Rename KILL_DISMISSED to KILL_RESET_KEEP_ITEMS and use it less
KILL_DISMISSED was identical to KILL_RESET except that non-summoned items
in the monster's inventory would be dropped, but whether one or the other
was used for a given function call was rather arbitrary (and it wasn't
intuitively clear how little difference there was without looking it up.)

This commit renames the enum value to be more explicit about what it does
and then changes most uses of KILL_DISMISSED to KILL_RESET instead. In the
overwhelming majority of cases this was either used for monsters that can't
drop items in the first place (ie: ioods, foxfires), or used after manually
setting MF_HARD_RESET on a monster (which is just a longer way to do the
same thing KILL_RESET does anyway). A few points in monster_die even
converted KILL_RESET to KILL_DISMISSED for specific enemies, like fire
vorticies for no reason I can determine (they can't have items, either!)

I have kept KILL_RESET_KEEP_ITEMS in the couple cases where preserving the
items is deliberate.

(It's *possible* that converting KILL_RESET to KILL_DISMISSED for dancing
weapons was intended to serve some actual function, but I'm not sure
there's any situation that should KILL_RESET an arbitrary monster and want
any piece of it to remain. If I've overlooked a good reason, that bit can
be changed back again.)