Reuse pillars of rime as occasional Cocytus decor
Another attempt to try and use what we already have available in game for
another decorative purpose, to make Cocytus feel a little more Hell like
instead of just being an ice zone with demons in it as ice caves already
present.

Compared to the usage of decorative boulders in d9500d9 and their
replacement for piles of debris in 03d5c24, these don't list a movespeed
or have use that itself is more active. (If people _are_ still going to be
paranoid about them doing something, then those same people are probably
still paranoid about granite statues or pillars of salt.) Also: by the
time one's in Cocytus, most characters should also be readily easily able
to break apart or pierce attack past an obstacle with only ~37 average HP
and 15 AC, so I am experimenting a second time with using obstacles that
don't start slowly dying on the slightest touch. It should distinguish it
a little more as an actual new decoration, and they're not really placed
in a branch that has lots of space or running around to encourage
individually destroying them all.