Remove (hopefully) unused pluralization code from summoned_poof_msg
While summoned_poof_msg was normally used for summoned monsters
disappearing, it could also be called with an item argument, which was used
when monsters opted to drop a summoned item (and that item subsequently
disappeared). The pluralization code was used only in this place, but I do
not believe it is actually possible for living monsters to drop summoned
items in a way the player can see it. (They sometimes do during their
creation, but this is silent, and they specifically can't pick up new gear
which might cause them to swap out their old gear.)

In the interest of simplification, I have removed this code and placed an
assert in monster::drop_item to catch the possibility that some case still
exists where a summoned item can become non-silently dropped.