More assorted tidying of monster_die
-Use a macro to more cleanly check for all 'reset-style' killer_types in
 some places.
-Move some cleanup-related code into monster_cleanup()
-Tweak what monster death effects are blocked by exploding (will no longer
 allow Gozag worshippers to make a dancing weapon actually fall to the
 ground by using Firestarter)
-Rename a confusingly-named and rarely used argument.
-Use the new 'real_death' boolean to shorten a common grouping of death
 checks in many on-kill effects.
-Add a small number of clarifying comments.

This isn't comprehensive. There's quite a few things which *probably*
should be in monster_cleanup() and aren't (which could matter in the
multiple situations where monster_cleanup() is called without
monster_die()). I also wanted to wrap a number of monster-specific death
effects into a function similar to _player_on_kill_effects() but there's
good reason for some of them to happen *before* death messaging and for
some of them to happen *afterward* (and while almost none of them trigger
for non-real_deaths, a couple need to).

But this still fixes several small bugs (and hopefully doesn't introduce as
many new ones!) while being slightly tidier.