Remove one-off handling of Awaken Vines
Awaken Vines' implementation used multiple single-use props, an
enchantment, and a custom removal function to manage what was *effectively*
just a normal summon cap and summon/summoner relationship. Especially with
the recent refactoring of summon handling to include non-abjurable
creations, this is entirely unnecessary, so let's cut it all.

(One bit of hard-to-even notice behavior that is not preserved by this
restructuring is that snaplasher vines would instanty die if the dryad
which created them ever lost LoS to them. This may have made some sense
when dryads were normal enemies, but that hasn't been the case in over a
decade.)