Fix Sniper showing 1% hit chance and bypass shield
Improvements to the hit chance display unfortunately introduced a bug
when hit chance is AUTOMATIC_HIT, causing the xv to show to hit as 1%.

While fixing this I noticed there is a very tiny but still non-zero
chance for AUTOMATIC_HIT attacks to be blocked by a shield, due to
AUTOMATIC_HIT being treated as a hit roll of 1500. Since this is either
slightly unexpected or if intended should be more impactful, it seems
easiest to just bypass shield checks as well for something like Sniper.