Slightly change find_habitable_spot_near and document
This adds the ability to set an 'exclude_radius' to the search, which will
ignore all spaces within that radius of the center point. This expands and
replaces the old 'allow_center' argument (since an exclude_radius of -1 is
equivalent to false, and 0 to true). It also generalizes the
'in_player_sight' argument to work for any actor, by passing a pointer to
that actor. Both of these will be used by a subsequent commit.

(I kind of dislike how radius and exclude_radius arguments aren't next to
each other, but since the latter rarely needs to be specified, I opted to
put it afterward so that it can be optional.)

The default behavior for this method is now equivalent to if allow_center
was true. Many uses of this method specified false, but I have not
preserved this behavior since in almost all cases it did nothing - the
center point in many searches is already occupied and thus will already be
ignored. The time savings of skipping a single cell check is irrelevant
compared to a simpler method call, imo.