Add new members to mgen_data to control placement radius
Prior to this commit, monsters created via create_monster() were either
created at a specified position (if the MG_FORCE_POS flag was set) or at a
random habitable spot within 2 tiles of a specified position (if it was
not). If a spell wanted to choose a random spot at any other distance, even
just '1', they needed to write their own code to do so - and there were
many examples already in the codebase of performing a very similar sort of
random iteration manually.

This commit aims to make that much easier to do, with less boilerplate, by
adding 3 new members to mgen_data: range_preferred, range_max, and
range_min.

ranged_preferred sets the range from the center point where the monster can
be randomly placed (ie: the current behavior functions as if this was
always '2')

range_max allows placement to exceed this range if, and only if, all tiles
within that range are invalid. This allows an easy fallback in cases where
we'd rather place something close, but will accept placing it far away
rather than not place it at all.

range_min specifies the minimum distance that the placement *must* be
further than - to allow forcing something to be placed non-adjacent to a
given location, for instance.

All of these can be conveniently set via mgen_data::set_range()

This commit also adds the MG_SEE_SUMMONER flag, which requires that the
chosen spot also be within line-of-sight of its summoner. (Most summon
spells do not require this, as they are centered around the summoner in the
first place, and line-of-sight to the center point is *always* required,
but some create summons at a random location around a target instead, where
this is useful.)

This commit also refactors multiple existing spells to benefit from this
new capacity - sometimes shortening them a great deal in the process.

There should be just a couple minor functional changes from this.
 -Dragon's Call can no longer create dragons outside of the player's LoS,
  if it chooses a target on the edge of the player's LoS
 -Haunt likewise will no longer place ghosts out of sight, when aimed at a
  distant enemy. This was awkward, since it was possible for it to appear
  as though the spell did nothing (and in practice, it might *actually* do
  nothing, since the ghosts can't attack from there anyway). I expect this
  is a small buff in practice, but I think it's just good policy.
 -Scroll of Butterflies uses trivially different placement logic for its
  more distant butterflies (but note that the edited comment was always
  inaccurate, due to a lack of the MG_FORCE_PLACE flag. It chose a random
  spot within 2 tiles of another random spot that was within 3 tiles,
  irrespective of whether the first chosen spot was empty, and thus always
  effectively picked a spot with 5 tiles. Since ?butterflies current
  behavior seems good to me, I've not elected to change the code to fit the
  comment, but rather the comment to fit the code.)