Placeholder feature: decorative floor
This is a bit of a placeholder hack for a large number of good tile
submissions in 76a730d and potential additional tile additions over time,
that don't want to actually be executed as rapidly-destructible or
diggable statues, solely placed on top of walls, or heavily written up for
character-specific messaging as fountains and food incidentally heavily
rely upon. (There are also no features left that entirely make sense for
them to be renames for, after fountains gained flavour messages in 8966e59
to help justify their whole lot. It would also add an indefinite number of
additional floor tiles to feign these otherwise being properly layered on
top of given branch layouts, and would lose out on writing up any
descriptions for them.)

Without any other relatively reasonable ways to approach such, and with
nearly a dozen different potential tiles workable for these purposes, I am
using a hack of a new feature solely meant to be redefined, which for all
intents and purposes is floor with a different glyph. (As with the
addition of food caches back in 831329d, these are all walkable features
that shouldn't interrupt terrain transformation in any fashion, and which
Xom doesn't interact with.)

Ideally, these cases would instead be handled with a significant rewrite
to the vault code for KFEAT to allow per-instance per-glyph redefinitions
of names and colours and tiles, in much the same way as vault monsters,
scythes, and hammers work. It'd deal with various other issues (like the
current inability to use both dead trees and live trees in the same vault,
or the inability to use more than two types of statues) and smoothen over
a fair bit of this code. Until then, the relatively invisible hack should
be worth the straightforward additional flavour possibilities.

(Note: I am 100% against adding almost anything else in this space, like
DNGN_METAL_STATUE_2 or DNGN_DECORATIVE_FLOOR_2, solely for the sake of
vaults that want to deploy multiples of these at once. Said rewrite should
happen before stretching this hack any further, as the cost of having a
single feature wouldn't be nearly as bad as adding further ones.)