Tweak the text of large HP costs on abilities
The previous ambigious cost of "Pain" was used on four different abilities:

 * Kiku's Brand Weapon With Pain: invoking Torment
 * Lugonu's Enter the Abyss: a completely random but non-lethal HP amount
 * Yred's Bind Soul: 1/3rd of HP, returned on soul-binding
 * Makhleb's Brand Self: 2/3rds of HP

Mentioning in abilities that this merely "will hurt you" is quite the
understatement, and was also entirely false for Kiku if the player is
undead or a wisp. "Pain" is also a relatively overloaded term in-game,
with both a weapon brand and a monster spell that both don't hurt nearly as
much.

As such, I've added another ability flag to differentiate Kiku's Torment
cost from the other costs, and more precisely state in the other ability's
cost descriptions that one will lose a large fraction of their HP.
Hopefully, this should prevent people from using Kiku and Makhleb 6* gifts
mid-fight as can be quite dangerous.