Player Charm: new effect for existing wand
Allow monsters to finally befriend the player.

Previously monsters could use the wand but would never use it on the
player; only on the player's allies. This made it useless against anyone
without allies, a generally a lower-tier threat especially compared to
how useful it is to the player.

A successful check will make the agent plus a number of additional
monsters in LOS (of the same species and/or band) gain a "charmer"
status, with a new pink halo to indicate the perceived alignment change.

You now believe these monsters are your friends, and any damage you
cause them (directly or indirectly) will cause such unbearable guilt
that you miss a turn. This is very similar to a single turn paralysis
(except it can affect even those normally immune to paralysis) leaving
you more prone to attack.

The new effect is designed to be a unique threat, especially serious on
a monster that has a pack of allies, but mechanically distinct from
other debuffs that tread similar "mind control" territory (e.g.
Mesmerisation or Fear)