Strolling Stone: new monster
A craggy, mossy beast inhabiting the early dungeon. When it attacks it
will shed plant-based detritus in the form of toadstools and lichen.

A mid-range addition to D that requires careful handling and frustrates
the player by littering the battlefield with obstacles.

The tile is a mash-up of the "orb statue", "guardian golem",
"stone_mossy2", with a dash of the old glowing mold.