Rock Fish: new wall monster
Rock Fish use the wall monster habitat to present a new threat in early
D that removes the advantage of 1-tile corridors as a defence and
introduces the mechanics of wall monsters to the player with a monster
that is otherwise quite vanilla.

It is usually found swimming in a small school to press the advantage
of being able to attack from more sides than a corridor would normally
allow.

Tile: a recoloured and tweaked lava fish.