Simplify monster choice of weapon for dual wielders
The monster::weapon() function contained two contradictory comments,
neither of which accurately described the actual logic taking place.

It seems that calling this method with no parameters, which_attack
defaults to -1 and in the case of a monster with the two_weapons flag
this meant a coinflip() would be used to decide whether main or offhand
weapons were returned.

This was certainly unexpected for a caller; you might expect the weapon
choice to be randomised during combat rounds but when which_attack was 0
or 1 it wouldn't randomise and would return main for 0 and offhand for
1. Values of 2 or above are seemingly never used for two_weapons
monsters who all have exactly 2 attacks anyway.

Looking at places that call this method, it could certainly have been
resulting in buggy behaviour, particularly for coglin player shadows, as
generally callers were expecting weapon() to return the same result
twice in a row if nothing had changed, which was not in fact guaranteed.

I removed the randomness entirely but if wanted we should randomly
reverse the *order* of weapons but still use both in a given dual wield
attack ... it surely never makes sense for them to fire the same weapon
twice in a turn.

Extracting the decision logic to another method I was also able to
dedupe the same logic when used for monster_info instead of monster;
which itself wasn't an accurate replica of the logic anyway.