Allow Okawaru gift reroll, at immense piety cost
Mechanically: add an ability that clears the existing set of gift
offerings, at a cost in piety that is guaranteed to drop standing below
6*. If neither gift has been accepted yet, both will be rerolled. If one
gift has been accepted, only the other is rerolled, and if both have
been accepted then this ability does not exist.

Motivation: This is a hybrid of old and new gifting behavior. The bad
playfeel of "unique, once-per-game capstone options are all garbage" is
removed. The player can choose to wait for something better, like in the
old system. However the gifts are still unique, and there is no
long-term piety-draining trickle of ignorable gifts once a capstone has
been selected.

Thematically: Rejecting Okawaru's offerings is a touchy topic to
navigate, and will cause offense, hence the large piety cost.

This is submitted as a POC conversation-starter. It is entirely untested.