New spell: Launch Clockwork Bee (level 3 Forgecraft)
A starter spell for Forgewrights (and their first that actually places a
monster on the field), this creates a fast and powerful-for-its-level
mechanical bee, with a number of specific limitations.

Casting the spell requires a hostile target to point it at, and then
winding the bee's clockwork requires channelling it for several turns
before it launches. The bee will fixate its attacks on the selected target
while it lives (similar to how Haunt and Soul Splinter work), but can
perform only a limited number of attacks before running out of power.

When the bee runs out of power, it falls dormant to the ground. The player
can wind it back up again by attempting to move into its tile, which will
refill it with energy and even heal it a little (and is a lot faster than
channelling 4 turns to launch a new one).

The idea is that the slow deployment (that cannot be done pre-battle),
combined with the strong incentive to actually poke your bee when it falls
down, will make playing with this ally a distinctive experience. And its
raw stats will make it feel good despite its limitations.

(It even comes with rF+/rElec+ to make it play a little better with
Forgewright's level 4 spells.)